# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Absher Travel & Tourism API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [flight, hotel, program, visa]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled]
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /dashboard/recent-bookings:
    get:
      operationId: getRecentBookings
      tags: [dashboard]
      summary: Get recent bookings for dashboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Recent bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"


  # ── AUTH ─────────────────────────────────────────────────────────────────
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already registered
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: New tokens issued
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, refreshToken]
                properties:
                  accessToken: { type: string }
                  refreshToken: { type: string }
  /auth/logout:
    post:
      operationId: logoutUser
      tags: [auth]
      security:
        - bearerAuth: []
      summary: Logout and revoke session
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: Logged out
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      security:
        - bearerAuth: []
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafeUser"

  # ── FLIGHTS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Booking:
      type: object
      required: [id, type, clientName, clientPhone, status, createdAt]
      properties:
        id: { type: integer }
        type: { type: string, enum: [flight, hotel, program, visa] }
        clientName: { type: string }
        clientPhone: { type: string }
        clientEmail: { type: ["string", "null"] }
        destination: { type: ["string", "null"] }
        travelDate: { type: ["string", "null"] }
        returnDate: { type: ["string", "null"] }
        adults: { type: integer }
        children: { type: integer }
        notes: { type: ["string", "null"] }
        status: { type: string, enum: [pending, confirmed, cancelled] }
        totalPrice: { type: ["number", "null"] }
        createdAt: { type: string }
    BookingInput:
      type: object
      required: [type, clientName, clientPhone]
      properties:
        type: { type: string, enum: [flight, hotel, program, visa] }
        clientName: { type: string }
        clientPhone: { type: string }
        clientEmail: { type: string }
        destination: { type: string }
        travelDate: { type: string }
        returnDate: { type: string }
        adults: { type: integer }
        children: { type: integer }
        notes: { type: string }
        totalPrice: { type: number }
    BookingUpdate:
      type: object
      properties:
        status: { type: string, enum: [pending, confirmed, cancelled] }
        notes: { type: string }
        totalPrice: { type: number }
...
    RegisterInput:
      type: object
      required: [password]
      properties:
        email: { type: string, format: email }
        phone: { type: string }
        password: { type: string, minLength: 8 }
        firstName: { type: string }
        lastName: { type: string }
        nationality: { type: string }
        gender:
          type: string
          enum: [male, female, other]
        dateOfBirth: { type: string, format: date }
    LoginInput:
      type: object
      required: [password]
      properties:
        email: { type: string, format: email }
        phone: { type: string }
        password: { type: string }
    SafeUser:
      type: object
      required: [id, role, isActive, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        firstName: { type: ["string", "null"] }
        lastName: { type: ["string", "null"] }
        nationality: { type: ["string", "null"] }
        gender: { type: ["string", "null"] }
        dateOfBirth: { type: ["string", "null"] }
        role: { type: string, enum: [customer, agent, admin, super_admin] }
        isActive: { type: boolean }
        emailVerifiedAt: { type: ["string", "null"] }
        phoneVerifiedAt: { type: ["string", "null"] }
        lastLoginAt: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    AuthResponse:
      type: object
      required: [user, accessToken, refreshToken]
      properties:
        user:
          $ref: "#/components/schemas/SafeUser"
        accessToken: { type: string }
        refreshToken: { type: string }

  # ── FLIGHT SCHEMAS ───────────────────────────────────────────────────────
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT